Loosen Cryptonite dependency upper bound
And fix bug I introduced when adding support for Cryptonite 0.14 in
a66f7222fe3f35cb86306b1e7f1a42eeda232cdf: I somehow managed to use the
number of hex chars (number of bits over 4) instead of the number of
bits in `params_p` when filling in the new `params_bits` field. By
coincidence the use `params_bits` in `i2ospOf` in Cryptonite was also
buggy in Cryptonite 0.14 but fixed in 0.15, and this resulted in the
spurious 0.19 upper bound (when 0.15 should have been the upper bound,
so I made another mistake there when I thought I had tested 0.18 but
had not).